Suppress Dialyzer errors (hopefully correctly)
Using `OK.for` produces a lot of these kinds of errors on lines that use the `<-` operator:

```
[ElixirLS Dialyzer] The variable _@6 can never match since previous clauses completely covered the type {'error',_} | {'ok',_}
```

According to the docs: 
> marks the given chunk as generated so it does not emit warnings. Currently it only works on special forms (for example, you can annotate a case but not an if).

I think adding this flag suppresses errors. After trying to make deliberate dialyzer errors within an `OK.for` block, I see that they are still getting picked up by dialyzer so I think adding this flag is safe.